+++ bo_forge/__init__.py
@@ -1,6 +1,6 @@
-"""BO Forge v2.1.0."""
+"""BO Forge v2.1.1."""
 
-__version__ = "2.1.0"
+__version__ = "2.1.1"
 
 from bo_forge.config import (
     BOConfig,